Encapsulated designer view into a separate module
Preparation for the viewer admin module
