+++ apps/gisday-competitions-angular/src/app/pages/admin/pages/designer/components/design-form/design-form.component.scss
