+++ apps/gisday-competitions-angular/src/app/pages/admin/pages/designer/components/design-question/design-question.component.spec.ts
