+++ apps/gisday-competitions-angular/src/app/pages/admin/pages/designer/designer.component.scss
