+++ apps/gisday-competitions-angular/src/app/pages/admin/pages/designer/designer.component.ts
@@ -3,10 +3,10 @@ import { Component } from '@angular/core';
 import { CompetitionForm } from '@tamu-gisc/gisday/competitions';
 
 @Component({
-  selector: 'tamu-gisc-admin',
-  templateUrl: './admin.component.html',
-  styleUrls: ['./admin.component.scss']
+  selector: 'tamu-gisc-designer',
+  templateUrl: './designer.component.html',
+  styleUrls: ['./designer.component.scss']
 })
-export class AdminComponent {
+export class DesignerComponent {
   public formModel: CompetitionForm;
 }
