+++ apps/gisday-competitions-angular/src/app/pages/admin/pages/designer/designer.module.ts
@@ -6,19 +6,20 @@ import { RouterModule, Routes } from '@angular/router';
 import { UIFormsModule } from '@tamu-gisc/ui-kits/ngx/forms';
 import { UILayoutModule } from '@tamu-gisc/ui-kits/ngx/layout';
 
-import { AdminComponent } from './admin.component';
+import { DesignerComponent } from './designer.component';
 import { DesignQuestionComponent } from './components/design-question/design-question.component';
 import { DesignFormComponent } from './components/design-form/design-form.component';
-import { FormsModule } from '../../modules/forms/forms.module';
+import { FormsModule } from '../../../../modules/forms/forms.module';
 
 const routes: Routes = [
   {
     path: '',
-    component: AdminComponent
+    component: DesignerComponent
   }
 ];
+
 @NgModule({
   imports: [CommonModule, RouterModule.forChild(routes), ReactiveFormsModule, UIFormsModule, UILayoutModule, FormsModule],
-  declarations: [AdminComponent, DesignQuestionComponent, DesignFormComponent]
+  declarations: [DesignerComponent, DesignQuestionComponent, DesignFormComponent]
 })
-export class AdminModule {}
+export class DesignerModule {}
